refactor(common): move BlockKey into pegaflow-common
