+++ pegaflow-common/src/lib.rs
@@ -1,7 +1,11 @@
+pub mod block;
 pub mod hll;
 pub mod logging;
+#[cfg(target_os = "linux")]
 pub mod numa;
 
+pub use block::BlockKey;
+#[cfg(target_os = "linux")]
 pub use numa::{
     NumaNode, NumaTopology, format_cpu_list, pin_thread_to_numa_node, query_pages_numa,
     read_cpu_topology_from_sysfs, run_on_numa,
